New statistics, without calling out to Plot.ly.
Operations should set a cron job to call `_maintenance_update_graph_data.sh`
periodically throughout the day. When there's a more permament server we'll
update the URL that script downloads from.

Right now, my server produces an update once per day, and that's all this
dataset has for resolution anyway, so it should be OK.

I have more datasets in progress, particularly around continents, countries, and
TLDs, but this conquers the first goal of fixing things to not be so suck.