Update the timeline TSV
There was an off-by-one error in the summation math for the numActive column.
Also, I've continued calculating old numActiveFQDNs and numActiveRegistereDomains
for the dataset, though it takes time for each row.

Updating this here is a little silly, since the maintenance task will update it
anyway, but eh, why not have the latest when this merges to master? :) And also,
I noticed the off-by-one error in the Github review, so it feels right to fix it.