Tighten CSP
Move Google Analytics to its own file and load it per guidance:

https://developers.google.com/analytics/devguides/collection/analyticsjs/#alternative_async_tracking_snippet

This in combination with PR #61 permits the following, much tighter CSP
policy:

> Content-Security-Policy "default-src 'self'; style-src 'unsafe-inline' 'self'; script-src 'unsafe-eval' 'self' https://www.google-analytics.com; img-src 'self' data: https://www.google-analytics.com;"

This can land whenever, but the CSP policy shouldn't be updated until #61 lands.
Additionally, the CSP policy should be updated with care and tested as well as
possible before being left alone.