Clarify and warn about staging roots.
In #112, we added a description of our staging roots to our main certificates
page. I think this is likely to lead to misunderstandings, and people adding
those certs to their trust stores.

In this change, I remove the description there, and expand on the description on
the staging-environment page to specifically warn against trusting those
certificates in non-testing environments.

I also removed the .txt and .der forms of those certificates, which I don't
think are necessary (also the .txt form didn't match the other .txt files on the
same page).