Run HTML Proofer on every CI build (#120)
This PR adds support to the travis build for running [HTML Proofer](https://github.com/gjtorikian/html-proofer) on each CI invocation. To do so requires adding a Gem to the Gemfile, creating a `rake test` Rake task to run the linter with a given configuration, and updating the travis yaml to run the rake task.

A [number of things](https://github.com/gjtorikian/html-proofer#whats-tested) are tested by `HTMLProofer`, but perhaps most useful is that it enforces that no links (internal or external) are 404ing at the time of build.

To get the linter to pass for the current website a number of fixes were made.
* Two sites with TLS misconfigurations (compose.com and certificateautomation.com) are ignored - I've contacted both out of band to troubleshoot the misconfiguration.
* Alt attributes were added to images that were missing them
* Client options page references to github.com/daplie were removed (they are 404s)
* The ISRG expired, revoked and valid X1 Root test webpages were excluded (they are deliberately not meant to validate with an older configuration)
* TLS 1.2 strict configuration websites were excluded (Need to work around the provided curl version). 
* The ALA website was ignored (It was timing out frequently)
* Crates.io is ignored (It returns JSON to curl out-of-box)

We can remove these exceptions as they are resolved & configured around. This is already a big patchset and I wanted feedback before it was perfect :-)