Tightens up CI for htmlproofer. (#136)
This commit removes most of the exceptions from the htmlproofer
Rakefile.

The failures that were fixed are in three categories:

1. Sites that timed-out with the base timeout. This was fixed by
   increasing the Typheous timeout to 60s
2. Sites that had chained up to Identrust's root CA
3. Sites that required TLS 1.2+

Items 2 and 3 were fixed by updating the project .travis.yml to add
`sudo: required`. Nothing in our CI pipeline truly requires `sudo` but
this flag is used to gate whether the Ruby build image is Ubuntu 12.04
or 14.04.

One exception remains for `crates.io`. The only workaround I could
identify to fix that false-positive created additional failures with
other sites so I gave up for now.