Updates Discourse's privacy policy URL.
The website CI failed due to a 500 error[0] accessing
`http://www.discourse.org/privacy-policy`. I was able to confirm this
issue and opened a thread[1] with Discourse who confirmed the URL
changed to `https://www.discourse.org/privacy`. They've added a redirect
but we might as well use the correct new URL :-)

[0]: https://travis-ci.org/letsencrypt/website/builds/219840564
[1]:
https://meta.discourse.org/t/discourse-org-privacy-policy-url-generates-500-internal-server-error/60616