Only give site description on homepage.
Previously we were setting a meta description tag for all pages that contained
the generic description "Let's Encrypt is a free, automated, and open certificate
authority brought to you by the non-profit Internet Security Research Group
(ISRG)." This gets inserted, e.g. into summaries of links pasted in
Discourse, which is confusing, because it makes it look as if the link goes to
the homepage.

In this change, only provide the default description for the homepage.