Removes `Daplie` from "Projects Integrating with LE" list.
The link to the "Daplie" project linked under "Projects Integrating with
LE" now 404s. Looking at the daplie.com website it's not clear what
happened to the "Walnut" or whether they still integrate with LE. I'm
going to remove the link for now and if someone can provide the correct
link and assurance the project still integrates with LE it can be
readded.