client-options.md: Fix acme.sh ACME v2 link
The ACME v2 link to ame.sh is broken. It was a link to the client's v2 branch; now the branch has been merged into master and deleted, so the link no longer works. Change it link to master.

Reported by rleeden on the forum. [1]

Unrelated: Remove an extra "/" from a different URL.

[1]: https://community.letsencrypt.org/t/acme-client-implementations-404-on-acme-sh-link/53596