Fix #107 Google Analytics: respect Do Not Track
Remove the dns-prefetch of www.google-analytics.com and www.googleadservices.com
Moved the loading of googletagmanager into ga.js
ga.js is now async and defer.
ga.js returns if Do Not Track is on.
gtag_report_conversion uses the href of getting-started-button (because the link may be different with i18n)