client-options: document inclusion/update policy.
Add instructions about sending a PR to update the client-options page.
Specifically mention where to add your client, and the trademark policy.

After internal discussion we've determined that browser based clients
do not provide the ACME experience we want to prioritize with our client
options page. This commit removes all of the browser based clients and
updates the page to describe this policy.