Please add 'FreeSSL.tech Auto' PHP client
Perhaps it is the first 3rd party client that does complete automation, including SSL certificate installation, in shared hosting cPanel.
No root access required.
It works with Live ACME V2 and V1 API. It is possible to install the client and configure with a browser. Developed for less tech-savvy users.