Show English blog posts on translated versions
This commit merges in missing blog post translations on the home page from the first site, i.e. English.

To get "first" to mean English, this commit also sets a distinct weight for each language to get the expected sort order (1. English 2. French 3. Spanish).

Note that the links to the blog articles in another language will go to the English version of the site, so you may want to make them open in another window to make that explicit.

Fixes #402