Mark the def shortcode as version 1
In the upcoming Hugo 0.55 we have adjusted how we handle the outer-most `{{%` delimiter. In short, it's now included when doing Markdown rendering and similar.

Which is normally a good thing, as it can partake in the ToC, footnotes etc.

The `def` shortcode, however, will not work with this setup, and is marked as `version: 1`. This setting will have no effect on older Hugo versions.