content: redirect to new jobs page
A search on Duck Duck Go for "letsencrypt jobs" still sends the user
to letsencrypt.org/jobs, which currently returns a 403 Forbidden.
That's not great so instead of returning a 403, redirect to the
current jobs page which is hosted at abetterinternet.org.

You can't do this with Hugo aliases because they are limited to
same-domain redirects; see gohugoio/hugo#5621 for more information.

For people who want to build the site without mucking with global NPM,
make it possible to install dependencies into node_modules/.bin and
load binaries from there in the build.sh script. Also clarify which
version of uglify needs to be installed (my initial reaction reading
"uglify-es" was that it was a typo but no that's the actual package
version that needs to be installed.)