Fix subscriber agreement link
The link in repository.md to the subscriber agreement points to
a document that doesn't exist. On the live website this was
"temporarily" fixed with a sym link. This change makes a copy in
the referenced location so the sym link will no longer be needed.
I have left the original document in place just in case anyone has
found that address, but it is not authoritative and has never been
referenced in the Let's Encrypt API, so I would not be opposed to
removing it (2017.11.15-LE-SA-v1.2.pdf).