Show only ACMEv2 clients on client-options.
Now that account creation is disabled, it doesn't make sense to
recommend a v1 client to any new users.

Tidied up the description of the Python acme module.

Removed the note to developers about submitting edits for upgrades to
v2. We have a footer with instructions on adding clients. I added an
ACMEv2 mention there.