Doc index: Add .html suffix to link to blog post
The link to the blog post about onboarding customers with Let's Encrypt
and ACME in the documentation index doesn't work using "hugo server -F"
(Hugo v0.59.1), but it works on the live site and on Netlify pull
request previews.

Add the missing .html suffix to the link in order to make it work
locally as well.