FAQ: New Google Safe Browsing URL
The URL https://www.google.com/safebrowsing/report_badware/ is used in
the FAQ but it redirects to
https://safebrowsing.google.com/safebrowsing/report_badware/ nowadays.
Change the URL to save the visitor a redirect.

Also escape the underscore in the URL since that's proper Markdown (vim
marks it with red otherwise).

Update all up-to-date translations to avoid unnecessary warnings about
out-of-date translation.