Link to PayPal Donate page in different languages
In order to show the PayPal Donate page in a particular language one has
to set two POST parameters:

* "lc" to either the full locale like en_US or just US (case
  insensitive)
* "country" to a country where that language is spoken or more
  specifically where PayPal supports that language (case insensitive)

Use the existing language codes for the "lc" parameter but introduce a
new variable paypalCountry for use in the "country" parameter and choose
one country per language.

Elaboration and discussions around this have taken place in issue #783.