Improve link shortcode and Glossary
The link shortcode now also accept the syntax `{{< link "/upcoming-features">}}` (without name for the link) which create a link with name the target url (with domain but without https://).

It improve the current syntax `[{{< relref "/upcoming-features" >}}]({{< relref "/upcoming-features" >}})` which is hard to read/understand.

Also fix last remaining relref (not found by rexepr because of a space between the name part and the link part of `[name](link)` : [name] (link)` in .md file and convert a ref into relref in an html file (for Netlify preview)