Update certificate hierarchy image for Dec 2020
The new version of the image shows the X3 and X4 intermediates as
retired, and correctly shows all end-entity issuance as coming from
R3 (as X3 is retired and E1 is not yet in use).

It also moves the image from /static/certs to /static/images, a much
more appropriate location. No translations need to be updated.