Cleaned up the Chain of Trust page
Made all wordings and references throughout the page fully consistent. Rewrote the cross-signing explanation to be clearer and remove redundant information already stated earlier on the page. Removed obsolete references, such as the OCSP signing certificate section that only applies to the completely-retired Let's Encrypt Authority line of intermediates. From my research, I do not believe that this signing certificate applies to R3 or R4 (and certainly doesn't apply to E1 or E2). Please confirm this.