Add sentence about searching on ct log monitors
We frequently see users on the Community just copy/pasting the expiry email. Usually I redirect them to this documentation page. It's almost always necessary to explain what happens using crt.sh.
I believe it would be a good thing if the documentation page I link to, also includes the part about ct log monitors as explained on the Community threads. That way one could not say the documentation is incomplete with regard to what happens in the Community threads.

This relatively simple addition refers to the use of ct log monitors to see a history of your certificates. This complements the part above it about adding/removing hostnames. Users might not remember doing that, so seeing the history on crt.sh is quite helpful I think and thus helpful to include it here.