Fix notbefore (#1384)
The CSV file and this page mention(ed) that the date in the CSV file is
the notBefore date of a certificate. However, it is the notAfter
datetime.

This is [already fixed in the
CSV](https://community.letsencrypt.org/t/questions-about-renewing-before-tls-alpn-01-revocations/170449/169),
but as [mentioned on the
Community](https://community.letsencrypt.org/t/is-the-date-in-the-third-column-notbefore-or-notafter/170912)
this page is still mentioning the date as notBefore (but in other
words).

This PR corrects the above.

This PR *also* changes back the `date` variable of the page to the date
the page was originally written, which was simultaneously updated with
the `lastmod` date in #1383 which *probably* shouldn't have. Correct me
if I'm wrong, but my gut says the `date` value is the date on which the
page was originally written and the `lastmod` value is the one which
needs to be updated with every update. And not `date`.