Adds renewc client to clients.json
My client [`renewc`](https://github.com/dvdsk/renewc) has reached the point where I am quite happy with it, so I am releasing it to the world now. I think it fits on this list.

It uses InstantACME (rust lib, on the list already) internally. The client compiles to a fully static binary that can be deployed without install (it has zero runtime dependencies not even a libc minimum version).

It provides an install command that copies the binary somewhere safe and sets up a systemd service and timer for renewal.

Renewal is checked every day (without api call). When the certificate gets close to expiry it is renewed. The point at which it does renewal is slightly random as required by letsencrypt.