Updated Getting Started content. (#1826)
The current Getting Started page is not nearly as helpful or clear as it
could be. We should simplify it, aiming the content squarely at the
kinds of people who might navigate to that page.

We should resist the temptation to over-complicate things. We don't need
to address every edge case on this page, and we can just assume that
anyone who has a chance of successfully getting and running an ACME
client on their own server already has basic sysadmin skills.

Also, we're not going to do a better job of explaining how to use
Certbot (or any other ACME client for that matter) than the Cerbot
website does, so let's just send people there for instructions on
operating a client.