Restructure Donate page to support translations by moving HTML to a layout file. (#1903)
Will fix #1898 

Also restores Donate item to nav for languages that have it previously.
For languages that do not have the word "Donate" translated, it does not
display in the menu.

When a language supports the Donate page, remove the `untranslated: 1`
property from that language's donate.html file.

CC @ludekjanda