Add new redirects for CP and CPS links (#2012)
This updates the stale subscriber agreement redirects, which exist as a
convenience for anyone who clicks a link that includes a trailing
period.

This adds new redirects from "https://cp-test.letsencrypt.org/" and
"https://cps-test.letsencrypt.org/" to the current version of the CPS.
Once we validate this works properly, we can stop using a standalone
webserver that hosts those URLs, and update this to use the real
domains.

DNS entries will be required for these to work, which will be handled
independently.