Add projects' git commit times
This commit adds data from a run of my Python script to collect the last
modification time of each project that has a readily identifiable git
URL. This data can be used for various purposes, most importantly hiding
clients on the Let's Encrypt web site that appear not to be actively
maintained. It's stored in Unix time format as a numeric field called
"last_commit".

Note that not all projects get annotated this way, because not all
of them have a git repository page (or one that can be determined
from the existing data).

A jq command line to view the clients with this field by their last
commit is

jq -r '."list"[] | select(."last_commit")? | "\(now - .last_commit | round): \(.name)" '  < clients.json

I'll separately commit the Python script itself somewhere so that it
can be rerun in the future to update this information.