Add Key Generation Report to Policy and Legal Repository (#2052)
This document asserts that the keys for Root YE and Root YR were
generated in accordance with the relevant requirements, and contains an
attestation from the auditors at BDO that said assertion is accurate. We
need to host this report from a static URL so that it can be included in
the CCADB metadata about these roots.