Docs for RFC6962 Logs EOL (#2099)
This updates the documentation on our ct-logs page to note the EOL of
our RFC6962 logs (and woodpecker along with it). I think we can include
more information here in the future, but this ensures we are at least
up-to-date with the state of the world.

Leave data/transparency.json in place for reference, though it's not
currently used.